Adapt literal expression tests
