+++ java/src/test/kotlin/io/spine/protodata/java/ExpressionSpec.kt
@@ -1,58 +1,35 @@
-///*
-// * Copyright 2023, TeamDev. All rights reserved.
-// *
-// * Licensed under the Apache License, Version 2.0 (the "License");
-// * you may not use this file except in compliance with the License.
-// * You may obtain a copy of the License at
-// *
-// * http://www.apache.org/licenses/LICENSE-2.0
-// *
-// * Redistribution and use in source and/or binary forms, with or without
-// * modification, must retain the above copyright notice and the following
-// * disclaimer.
-// *
-// * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-// * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-// * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-// * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-// * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-// * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-// * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-// * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-// */
-//
-//package io.spine.protodata.java
-//
-//import assertCode
-//import com.google.protobuf.ByteString
-//import io.spine.protobuf.TypeConverter
-//import org.junit.jupiter.api.DisplayName
-//import org.junit.jupiter.api.Nested
-//import org.junit.jupiter.api.Test
-//
-//@DisplayName("`Expression` should")
-//internal class ExpressionSpec {
-//
-//    @Nested
-//    inner class Print {
-//
-//        @Test
-//        fun `Java string literal`() = assertCode(LiteralString("foo"), "\"foo\"")
-//
-//        @Test
-//        fun `a number`() = assertCode(Literal(42), "42")
-//
-//        @Test
-//        fun `a boolean`() = assertCode(Literal(false), "false")
-//
-//        @Test
-//        fun anything() {
-//            val anything = "Frankie says relax"
-//            val expression = Literal(anything)
-//            assertCode(expression, anything)
-//        }
-//    }
-//}
+package io.spine.protodata.java
+
+import io.kotest.matchers.shouldBe
+import io.kotest.matchers.shouldNotBe
+import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Test
+
+@DisplayName("`Expression` should")
+internal class ExpressionSpec {
+
+    private val eight = "4 + 4"
+
+    @Test
+    fun `return the provided code`() {
+        Expression<Int>(eight).code shouldBe eight
+    }
+
+    @Test
+    fun `return code as string representation`() {
+        "${Expression<Int>(eight)}" shouldBe eight
+    }
+
+    @Test
+    fun `compare two expressions`() {
+        val expression = "2 + 2 * 2"
+        val six = Expression<Int>(expression)
+        val anotherSix = Expression<Int>(expression)
+        six shouldBe anotherSix
+        six.hashCode() shouldBe anotherSix.hashCode()
+
+        val nine = Expression<Int>("9")
+        six shouldNotBe nine
+        six.hashCode() shouldNotBe nine.hashCode()
+    }
+}
